+++ R/convertPedigree.R
@@ -2,24 +2,28 @@
 #' @param ped a pedigree dataset.  Needs ID, momID, and dadID columns
 #' @param component character.  Which component of the pedigree to return.  See Details.
 #' @param max.gen the maximum number of generations to compute
-#'  (e.g., only up to 4th degree relatives).  The default of Inf uses as many
-#'  generations as there are in the data.
+#'  (e.g., only up to 4th degree relatives).  The default is 25. However it can be set to infinity.
+#'   `Inf` uses as many generations as there are in the data.
 #' @param sparse logical.  If TRUE, use and return sparse matrices from Matrix package
 #' @param verbose logical.  If TRUE, print progress through stages of algorithm
 #' @param gc logical. If TRUE, do frequent garbage collection via \code{\link{gc}} to save memory
 #' @param flatten.diag logical. If TRUE, overwrite the diagonal of the final relatedness matrix with ones
 #' @param standardize.colnames logical. If TRUE, standardize the column names of the pedigree dataset
+#' @param tcross.alt.crossprod logical. If TRUE, use alternative method of using Crossprod function for computing the transpose
+#' @param tcross.alt.star logical. If TRUE, use alternative method of using \%\*\% for computing the transpose
 #' @param ... additional arguments to be passed to \code{\link{ped2com}}
-#' @details The algorithms and methodologies used in this function are further discussed and exemplified in the vignette titled "examplePedigreeFunctions".
+#' @details The algorithms and methodologies used in this function are further discussed and exemplified in the vignette titled "examplePedigreeFunctions". For more advanced scenarios and detailed explanations, consult this vignette.
 #' @export
 #'
 ped2com <- function(ped, component,
-                    max.gen = Inf,
+                    max.gen = 25,
                     sparse = FALSE,
                     verbose = FALSE,
                     gc = FALSE,
                     flatten.diag = FALSE,
                     standardize.colnames = TRUE,
+                    tcross.alt.crossprod = FALSE,
+                    tcross.alt.star = FALSE,
                     ...) {
   # Validate the 'component' argument and match it against predefined choices
   component <- match.arg(tolower(component),
@@ -152,8 +156,6 @@ ped2com <- function(ped, component,
   # compute rel <- tcrossprod(rsq)
   if (gc) {
     rm(isPar, newIsPar)
-  }
-  if (gc) {
     gc()
   }
   if (verbose) {
@@ -162,14 +164,24 @@ ped2com <- function(ped, component,
   r2 <- r %*% Matrix::Diagonal(x = sqrt(isChild), n = nr)
   if (gc) {
     rm(r, isChild)
-  }
-  if (gc) {
     gc()
   }
   if (verbose) {
     cat("Doing tcrossprod\n")
   }
-  r <- Matrix::tcrossprod(r2)
+  if (tcross.alt.crossprod) {
+    if (verbose) {
+      cat("Doing alt tcrossprod crossprod t \n")
+    }
+    r <- crossprod(t(as.matrix(r2)))
+  } else if (tcross.alt.star) {
+    if (verbose) {
+      cat("Doing alt tcrossprod %*% t \n")
+    }
+    r <- r2 %*% t(as.matrix(r2))
+  } else {
+    r <- Matrix::tcrossprod(r2)
+  }
   if (component == "generation") {
     return(gen)
   } else {
@@ -189,12 +201,11 @@ ped2com <- function(ped, component,
 
 #' Take a pedigree and turn it into an additive genetics relatedness matrix
 #' @inheritParams ped2com
-#' @details The algorithms and methodologies used in this function are further discussed and exemplified in the vignette titled "examplePedigreeFunctions".
-#' For more advanced scenarios and detailed explanations, consult this vignette.
-
+#' @inherit ped2com details
 #' @export
 #'
-ped2add <- function(ped, max.gen = Inf, sparse = FALSE, verbose = FALSE, gc = FALSE, flatten.diag = FALSE) {
+ped2add <- function(ped, max.gen = 25, sparse = FALSE, verbose = FALSE, gc = FALSE, flatten.diag = FALSE, standardize.colnames = TRUE,
+                    tcross.alt.crossprod = FALSE, tcross.alt.star = FALSE) {
   ped2com(
     ped = ped,
     max.gen = max.gen,
@@ -202,17 +213,20 @@ ped2add <- function(ped, max.gen = Inf, sparse = FALSE, verbose = FALSE, gc = FA
     verbose = verbose,
     gc = gc,
     component = "additive",
-    flatten.diag = flatten.diag
+    flatten.diag = flatten.diag,
+    standardize.colnames = standardize.colnames,
+    tcross.alt.crossprod = tcross.alt.crossprod,
+    tcross.alt.star = tcross.alt.star
   )
 }
 
 #' Take a pedigree and turn it into a mitochondrial relatedness matrix
 #' @inheritParams ped2com
-#' @details The algorithms and methodologies used in this function are further discussed and exemplified in the vignette titled "examplePedigreeFunctions".
+#' @inherit ped2com details
 #' @export
 #' @aliases ped2mt
 #'
-ped2mit <- ped2mt <- function(ped, max.gen = Inf, sparse = FALSE, verbose = FALSE, gc = FALSE, flatten.diag = FALSE) {
+ped2mit <- ped2mt <- function(ped, max.gen = 25, sparse = FALSE, verbose = FALSE, gc = FALSE, flatten.diag = FALSE, standardize.colnames = TRUE, tcross.alt.crossprod = FALSE, tcross.alt.star = FALSE) {
   ped2com(
     ped = ped,
     max.gen = max.gen,
@@ -220,19 +234,19 @@ ped2mit <- ped2mt <- function(ped, max.gen = Inf, sparse = FALSE, verbose = FALS
     verbose = verbose,
     gc = gc,
     component = "mitochondrial",
-    flatten.diag = flatten.diag
+    flatten.diag = flatten.diag,
+    standardize.colnames = standardize.colnames,
+    tcross.alt.crossprod = tcross.alt.crossprod,
+    tcross.alt.star = tcross.alt.star
   )
 }
 
-
-
-
 #' Take a pedigree and turn it into a common nuclear environmental relatedness matrix
 #' @inheritParams ped2com
-#' @details The algorithms and methodologies used in this function are further discussed and exemplified in the vignette titled "examplePedigreeFunctions".
+#' @inherit ped2com details
 #' @export
 #'
-ped2cn <- function(ped, max.gen = Inf, sparse = FALSE, verbose = FALSE, gc = FALSE, flatten.diag = FALSE) {
+ped2cn <- function(ped, max.gen = 25, sparse = FALSE, verbose = FALSE, gc = FALSE, flatten.diag = FALSE, standardize.colnames = TRUE, tcross.alt.crossprod = FALSE, tcross.alt.star = FALSE) {
   ped2com(
     ped = ped,
     max.gen = max.gen,
@@ -240,13 +254,16 @@ ped2cn <- function(ped, max.gen = Inf, sparse = FALSE, verbose = FALSE, gc = FAL
     verbose = verbose,
     gc = gc,
     component = "common nuclear",
-    flatten.diag = flatten.diag
+    flatten.diag = flatten.diag,
+    standardize.colnames = standardize.colnames,
+    tcross.alt.crossprod = tcross.alt.crossprod,
+    tcross.alt.star = tcross.alt.star
   )
 }
 
 #' Take a pedigree and turn it into an extended environmental relatedness matrix
 #' @inheritParams ped2com
-#' @details The algorithms and methodologies used in this function are further discussed and exemplified in the vignette titled "examplePedigreeFunctions".
+#' @inherit ped2com details
 #' @export
 #'
 ped2ce <- function(ped) {